requirement.txt
#for .env
python-dotenv
requests
ruamel.yaml